requirements.txt
requests==2.28.*
environs==9.5.*
python-telegram-bot==13.*